requirements.txt
asgiref==3.7.2
attrs==23.1.0
authorizenet==1.1.4
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
colander==2.0
Django==4.2.1
django-admin-numeric-filter==0.1.9
django-admin-rangefilter==0.10.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-countries==7.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-spectacular==0.26.3
idna==3.4
inflection==0.5.1
iso8601==2.0.0
jsonschema==4.17.3
lxml==4.9.2
Pillow==9.5.0
pycodestyle==2.10.0
pyrsistent==0.19.3
pytz==2023.3
PyXB==1.2.5
PyYAML==6.0
reportlab==4.0.4
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
translationstring==1.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
